% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ACCBasketball}
\alias{ACCBasketball}
\title{NCAA Men's Basketball data}
\format{Two matrices and a data frame:
\describe{
  \item{adjacencyMatrix}{Matrix in which the (i,j) element is the number of points, in thousands, that team i scored on team j over the regular season.}
  \item{covariateData}{Data frame that includes that name of the school and its enrollment.}
  \item{nGames}{Matrix in which the (i,j) element is the number of games played between teams i and j.}
}}
\source{
\url{spreadsheetsports.com}
}
\usage{
data(ACCBasketball)
}
\description{
This directed network of 15 nodes represents scoring between
teams in the Atlantic Coast Conference in the 2016 Regular season.
}
\references{
Gathered from spreadsheetsports.com.
}
\keyword{datasets}
