% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{authorship_data}
\alias{authorship_data}
\title{Internal. Used to create \code{authorship_data_1} and \code{authorship_data_3} datasets.}
\usage{
authorship_data(file_path, save = F, file = NA)
}
\arguments{
\item{file_path}{(character) The path to the raw scraped .csv data.}

\item{save}{(bool)}

\item{file}{(character)}
}
\description{
Takes as input a .csv file provided by TescPerf Python library, collected
from pure.uvt.nl, and outputs a list of collaboration. Each of the
collaborations in the list represents a paper and the TSB members
that collaborated on that paper. The order in the vector is
given by the order of occurrence in the list scrapped from
pure.uvt.nl.
}
