% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{authorship_matrix}
\alias{authorship_matrix}
\title{Internal. Used to build \code{authorship_matrix_1} and \code{authorship_matrix_3} datasets.}
\usage{
authorship_matrix(authorship_data, verbose = T)
}
\arguments{
\item{authorship_data}{(list of vectors) Each vectors represents the TSB authors that collaborated on that paper.}

\item{verbose}{(bool)}
}
\value{
a square matrix with the TSB collaborators and the subsequent weights indicating the way they collaborated.
}
\description{
Builds the matrix of weights for all the authors in the TSB at a point in time.
The first one is considered to have collaborated with the rest
of the authors in the vector. Each such edge is awarded
one point in the matrix of weights.
}
