% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Books}
\alias{Books}
\title{Books about US Politics}
\format{An undirected statnet network onject with 105 nodes
\describe{
\item{title}{Title of the book.}
\item{political_leaning}{The political leaning of the book, as compiled
    by Mark Newman from descriptions on Amazon. These have been recoded
    to "liberal", "neutral", and "conservative".}
}}
\source{
\url{http://vlado.fmf.uni-lj.si/pub/networks/data/mix/mixed.htm}
}
\usage{
data(Books)
}
\description{
This network consistens of 105 vertices, each representing a book about
politics sold in the early 2000s in Amazon.
}
\details{
Newman's note on the data:

``Books about US politics \newline
Compiled by Valdis Krebs

Nodes represent books about US politics sold by the online bookseller
Amazon.com.  Edges represent frequent co-purchasing of books by the same
buyers, as indicated by the "customers who bought this book also bought
these other books" feature on Amazon.

Nodes have been given values "l", "n", or "c" to indicate whether they are
"liberal", "neutral", or "conservative".  These alignments were assigned
separately by Mark Newman based on a reading of the descriptions and
reviews of the books posted on Amazon.

These data should be cited as V. Krebs, unpublished,
\url{http://www.orgnet.com/}.''
}
\references{
\url{http://www-personal.umich.edu/~mejn/netdata/};
    Valdis Krebs, Unpublished.
}
\keyword{datasets}
