% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{Colors}
\alias{Colors}
\title{Colors used in Inferential Network Analysis}
\usage{
Colors(classes = 3)
}
\arguments{
\item{classes}{Number of classes between 1 and 6 in the color vector.}
}
\value{
A vector of colors.
}
\description{
Returns a vector of colors to be used in plotting.
}
\details{
Returns a vector of colors (hexadecimal representation), which are sourced
from the http://colorbrewer2.org website.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
