% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estab.R
\name{estab}
\alias{estab}
\title{Bioentity establishment or not at each node}
\usage{
estab(
  decvec,
  dispvec,
  probestabmean4,
  probestabsd4,
  maneffdir4,
  maneffmean4n,
  maneffsd4n,
  plotmp = F,
  geocoords4n,
  readprobestabvec4,
  probestabvec4
)
}
\arguments{
\item{decvec}{vector of 1=management, 0=no management}

\item{dispvec}{vector of 1=sp has dispersed to node, 0=sp has not dispersed to node}

\item{probestabmean4}{mean probability of establishment (new or CONTINUED) in absence of management (used if \code{readprobestabvec4} =  F)}

\item{probestabsd4}{sd in probability of establishment in absence of management (used if \code{readprobestabvec4} = F)}

\item{maneffdir4}{if maneffdir4='decrease_estab', the management reduces the probability of establishment; if maneffdir4='increase_estab', the management reduces the probability that establishment does NOT occur}

\item{maneffmean4n}{mean effect of management (proportional change in estabp, where the combination of maneffmean4n = 1 and maneffsd4n = 0 makes establishment impossible)}

\item{maneffsd4n}{sd of management effect}

\item{plotmp}{if T, then map of establishment is plotted}

\item{geocoords4n}{matrix of x,y coordinates of nodes (used if \code{plotmp} = T)}

\item{readprobestabvec4}{if T, then a vector \code{probestabvec4} is read in; otherwise the vector is generated using \code{probestabmean4} and \code{probestabsd4}}

\item{probestabvec4}{vector of probabilities of establishment (read in or generated when the function is run, depending on \code{readprobestabvec4} equal to T or F)}
}
\description{
This function determines whether a bioentity establishes or not at each node.  It is similar in some ways to the function \code{makedec}, which determines whether management is adopted.
}
\details{
Updated 2020-09-05
}
\examples{
x6decrease <- estab(decvec=c(1,1,1,0,0,0), dispvec=c(1,1,1,1,1,0), geocoords4n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T, ncol=2), readprobestabvec4=F, probestabmean4=0.5, probestabsd4=0.1, maneffdir4='decrease_estab', maneffmean4n=0.1, maneffsd4n=0.1, plotmp=T)
x7decrease <- estab(decvec=c(1,1,1,0,0,0), dispvec=c(1,1,1,1,1,0), geocoords4n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T, ncol=2), readprobestabvec4=F, probestabmean4=0.5, probestabsd4=0.1, maneffdir4='decrease_estab', maneffmean4n=0.9, maneffsd4n=0.1, plotmp=T)
x8decrease <- estab(decvec=c(1,1,1,0,0,0), dispvec=c(1,1,1,1,1,0), geocoords4n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T, ncol=2), readprobestabvec4=T, probestabvec4=c(0.9,0.1,0.1,0.1,0.1,0.9), maneffdir4='decrease_estab', maneffmean4n=0.9, maneffsd4n=0.1, plotmp=T)
x9decrease <- estab(decvec=c(1,1,1,0,0,0), dispvec=c(1,1,1,1,1,0), geocoords4n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T, ncol=2), readprobestabvec4=T, probestabvec4=c(0.9,0.1,0.1,0.1,0.1,0.9), maneffdir4='decrease_estab', maneffmean4n=0.1, maneffsd4n=0.1, plotmp=T)
x6increase <- estab(decvec=c(1,1,1,0,0,0), dispvec=c(1,1,1,1,1,0), geocoords4n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T, ncol=2), readprobestabvec4=F, probestabmean4=0.5, probestabsd4=0.1, maneffdir4='increase_estab', maneffmean4n=0.1, maneffsd4n=0.1, plotmp=T)
x7increase <- estab(decvec=c(1,1,1,0,0,0), dispvec=c(1,1,1,1,1,0), geocoords4n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T, ncol=2), readprobestabvec4=F, probestabmean4=0.5, probestabsd4=0.1, maneffdir4='increase_estab', maneffmean4n=0.9, maneffsd4n=0.1, plotmp=T)
x8increase <- estab(decvec=c(1,1,1,0,0,0), dispvec=c(1,1,1,1,1,0), geocoords4n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T, ncol=2), readprobestabvec4=T, probestabvec4=c(0.9,0.1,0.1,0.1,0.1,0.9), maneffdir4='increase_estab', maneffmean4n=0.9, maneffsd4n=0.1, plotmp=T)
x9increase <- estab(decvec=c(1,1,1,0,0,0), dispvec=c(1,1,1,1,1,0), geocoords4n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T, ncol=2), readprobestabvec4=T, probestabvec4=c(0.9,0.1,0.1,0.1,0.1,0.9), maneffdir4='increase_estab', maneffmean4n=0.1, maneffsd4n=0.1, plotmp=T)
}
\keyword{establishment}
