% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_tesc_authors}
\alias{extract_tesc_authors}
\title{Performs the extraction of TESC authors from a larger list of authors.}
\usage{
extract_tesc_authors(authors)
}
\arguments{
\item{authors}{(vector) The character vector from which the TESC authors are extracted.
Typically it is the TSB vector of authors obtained from \code{rownames(authorship_matrix_1)} or
\code{rownames(authorship_matrix_3)}.}
}
\value{
A vector with the indices where TESC authors are found.
}
\description{
Can be useful to extract only the TESC authors from the TSB vector of authors.
The TESC authors vector is based on the \code{tesc_data} data frame.
See \code{\link{load_ina_data}} function for more details about the datasets.
}
