% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{find_name}
\alias{find_name}
\title{Internal. Find the location of a name inside a vector.}
\usage{
find_name(name, inside)
}
\arguments{
\item{name}{(character) The name that is used for the search.}

\item{inside}{(vector) The character vector where the \code{name} is searched in.}
}
\value{
The index where the match occurred or \code{integer(0)} otherwise.
}
\description{
Internal. Find the location of a name inside a vector.
}
