% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaoh}
\alias{gaoh}
\title{Golden Age of Hollywood Actor Networks}
\format{Seven network objects and one data frame:
\describe{
\item{n1--n7}{Each of these seven objects is is a network object.
    Each node is an actor, and an edge from i to j indicates that
    i appeard in at least one film in a supporting role to j during
    the respective time period.}
\item{attributes}{A data frame that includes vertex attributes.}}}
\usage{
data(gaoh)
}
\description{
This is a directed network of film actors with 55 nodes, over seven time periods
}
\references{
Taylor, Dane, Sean A. Myers, Aaron Clauset, Mason A. Porter,
 and Peter J. Mucha. "Eigenvector-based centrality measures for temporal
 networks." Multiscale Modeling & Simulation 15, no. 1 (2017): 537-574.
}
\keyword{datasets}
