% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genlocs.R
\name{genlocs}
\alias{genlocs}
\title{Generate geographic locations of nodes in network}
\usage{
genlocs(xrange4, yrange4, numnodes4, randgeo4 = TRUE)
}
\arguments{
\item{xrange4}{range of x coordinates}

\item{yrange4}{range of y coordinates}

\item{numnodes4}{the number of nodes}

\item{randgeo4}{if TRUE then locations are randomly generated (other options for distributing locations will be added)}
}
\description{
This function generates the geographic locations of nodes in network, for simulation studies where geographic locations are not observed.
}
\details{
Updated 2020-09-05
}
\examples{
geolocs99 <- genlocs(xrange4=c(0,50), yrange4=c(0,50), numnodes4=100, randgeo4=TRUE)
}
\keyword{geography}
\keyword{locations}
