% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_all_last_names}
\alias{get_all_last_names}
\title{Internal. Applies get_last_name function to a list of vectors (paper collaboration).}
\usage{
get_all_last_names(list_authors)
}
\arguments{
\item{list_authors}{(list of vectors)}
}
\value{
A new list of vectors containing only the last names.
}
\description{
Internal. Applies get_last_name function to a list of vectors (paper collaboration).
}
