% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_last_name}
\alias{get_last_name}
\title{Internal. Gets the last word of a full name (i.e., the last name). E.g., M. A. Constantin -> Constantin.}
\usage{
get_last_name(full_name)
}
\arguments{
\item{full_name}{(string or vector)}
}
\description{
Tried against these scenarios: D.M. Beneken Genaamd Kolmer, M.J. Brandt, Geert H. van Kollenburg, Jac J. L. van der Klink, L.V. van de Poll-Franse,
Y.T.M. Vanneste-van Zandvoort, M.A.L.M. vanAssen, L.V. van de PollFranse, L.V. van de PFranse, E.A.L. De Caluw? (? = special character).
}
\details{
The regex used: \code{([A-Z\\u00C0-\\u017F][a-z\\u00C0-\\u017F]+-)?([A-Z\\u00C0-\\u017F][a-z\\u00C0-\\u017F]+)$}
}
