% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hookups}
\alias{Hookups}
\title{Grey's Anatomy hookup network}
\format{A undirected statnet network object with 44 nodes:
\describe{
  \item{birthyear}{}
  \item{name}{}
  \item{position}{}
  \item{race}{}
  \item{season}{}
  \item{sex}{}
  \item{sign}{}
  \item{vertex.names}{}
}}
\source{
\url{}
}
\usage{
data(Hookups)
}
\description{
This undirected network consists of 44 vertices representing
}
\references{
Gathered by Gary Weissman and Augmented by Benjamin Lind
}
\keyword{datasets}
