% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{ina}
\alias{ina}
\alias{ina-package}
\title{Internship on Network Analysis for authorship data}
\description{
This package was created as a product of my internal internship at Tilburg University
on the subject of Network Analysis for TSB/ TESC authorship data.
}
\details{
To get started type \code{\link{load_ina_data}}. This will load the datasets. For a better
understanding of the differences between the datasets loaded be sure to check the
documentation for \code{\link{load_ina_data}}. Have fun!
}
