% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initvals.R
\name{initvals}
\alias{initvals}
\title{Generate initial status for nodes}
\usage{
initvals(geocoords4s, dist4, init.p4, init.n4, norp4, plotmp = F)
}
\arguments{
\item{geocoords4s}{n by 2 matrix of geographic coordinates, provided for a particular system or generated by function genlocs}

\item{dist4}{the type of locations where initial presence occurs: 'random' indicates all equally likely, 'upedge' indicates that nodes closest to the upper edge have presence, 'rightedge' indicates that nodes closest to the right edge have presence}

\item{init.p4}{the proportion of initial locations for presence}

\item{init.n4}{the number of initial locations for presence}

\item{norp4}{'num' indicates initial number for presence, 'prop' indicates initial proportion}

\item{plotmp}{if T, then map of presence and absence is plotted}
}
\description{
This function generates the initial status for nodes, in terms of the presence of management information or focal species
}
\details{
Updated 2020-09-05
}
\examples{
x2 <- initvals(j2 <- genlocs(xrange4=c(0,50), yrange4=c(0,50), numnodes4=100, randgeo4=TRUE),dist4='random', init.p4=0.1, norp4='prop', plotmp=T)
x3 <- initvals(j2 <- genlocs(xrange4=c(0,50), yrange4=c(0,50), numnodes4=100, randgeo4=TRUE),dist4='upedge', init.p4=0.3, norp4='prop', plotmp=T)
x4 <- initvals(j2 <- genlocs(xrange4=c(0,50), yrange4=c(0,50), numnodes4=100, randgeo4=TRUE), dist4='rightedge', init.p4=0.3, norp4='prop',plotmp=T)
x5 <- initvals(geocoords4s=matrix(c(1,5, 2,5, 3,5, 4,5), byrow=T, ncol=2), dist4='upedge', norp4='num', init.n4=2, plotmp=T) 
x6 <- initvals(geocoords4s=matrix(c(5,1, 5,2, 5,3, 5,4), byrow=T, ncol=2), dist4='rightedge', norp4='num', init.n4=2, plotmp=T) 
}
\keyword{conditions}
\keyword{initial}
