% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is_empty}
\alias{is_empty}
\title{Internal. Determines whether a result is of type integer(0) or character(0).}
\usage{
is_empty(input)
}
\arguments{
\item{input}{(any)}
}
\description{
Internal. Determines whether a result is of type integer(0) or character(0).
}
