% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerplot.R
\name{layerplot}
\alias{layerplot}
\title{Function for plotting multilayer networks}
\usage{
layerplot(
  upmat,
  lowmat,
  upcoords,
  lowcoords,
  yflat2,
  xover2,
  htdif,
  lwdhere = 1
)
}
\arguments{
\item{upmat}{adjacency matrix for upper network}

\item{lowmat}{adjacency matrix for lower network}

\item{upcoords}{x-y coordinates for upper network}

\item{lowcoords}{x-y coordinates for lower network}

\item{yflat2}{proportional flattening of map}

\item{xover2}{proportional right movement of map}

\item{htdif2}{distance between upper and lower network}
}
\description{
This function plots a multilayer network.  It assumes that the x and y coordinates are in [0,1]
}
\examples{
y1 <- matrix(c(0,0,0,1, 0,0,1,0, 0,0,0,1, 0,0,0,0),byrow=T, ncol=4)
y2 <- matrix(c(0,1,0,0,0, 0,0,1,0,0, 0,0,0,0,0, 0,0,0,0,0, 0,0,0,1,0),byrow=T, ncol=5)
x1 <- matrix(c(0.5,0.8, 0.9,0.9, 0.1,0.1, 0.5,0.5), byrow=T, ncol=2)
x2 <- matrix(c(0.5,0.8, 0.9,0.9, 0.1,0.1, 0.5,0.5, 0.3,0.8), byrow=T, ncol=2)
layerplot(upmat=y1,lowmat=y2,upcoords=x1,lowcoords=x2,yflat=0.5,xover=0.5, htdif=1)
}
\keyword{multilayer}
\keyword{networks}
\keyword{plotting}
