% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentnet.R
\name{loc}
\alias{loc}
\alias{loc.ergmm}
\title{Extract Estimated Locations from Latent Space Model}
\usage{
loc(object, ...)

\method{loc}{ergmm}(object, ...)
}
\arguments{
\item{object}{Model object returned by \code{\link[latentnet]{ergmm}}.}

\item{...}{Additional arguments, unused.}
}
\value{
Matrix of posterior means of the vertices as estimated in the latent
    space model.
}
\description{
Extracts the estimated posterior means for the vertices as estimated in the
\code{\link{latentnet}} package.
}
\details{
\code{loc} returns a matrix of posterior means for the latent locations of
each vertex as estimated by \code{\link[latentnet]{ergmm}}.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
