% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentnet.R
\name{mcmc_distance}
\alias{mcmc_distance}
\alias{mcmc_distance.ergmm}
\title{Extract Posterior Samples of the Distances Between Two Vertices}
\usage{
mcmc_distance(object, i, j)

\method{mcmc_distance}{ergmm}(object, i, j)
}
\arguments{
\item{object}{\code{\link[latentnet]{ergmm}} model object}

\item{i}{integer, representing the index of the first vertex}

\item{j}{integer, representing the index of the second vertex}
}
\value{
Returns an \code{\link[coda]{mcmc}} object
}
\description{
Extract the postertior samples of the distances between two vertices as
estimated in a latent space model.
}
\details{
\code{mcmc_distance} returns an \code{\link[coda]{mcmc}} object containing
the distances between the two specified vertices.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
