% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentnet.R
\name{model_performance}
\alias{model_performance}
\alias{model_performance.ergmm}
\title{Return the Model Performance of an Latent Space Model}
\usage{
model_performance(object, measure, x.measure = "cutoff",
  type = "pmean", ...)

\method{model_performance}{ergmm}(object, measure, x.measure = "cutoff",
  type = "pmean", ...)
}
\arguments{
\item{object}{\code{\link[latentnet]{ergmm}} model object}

\item{measure}{Performance measure to use for the evaluation. See
\code{\link[ROCR]{performance}}.}

\item{x.measure}{A second performance measure. See
\code{\link[ROCR]{performance}}.}

\item{type}{Prediction type to pass to \code{\link{predict.ergmm}}. Unlike
in that function, "pmean" is the default.}

\item{...}{Additional arguments, not used}
}
\value{
A \code{\link[ROCR]{performance}} object.
}
\description{
Returns a performance object for a model estimates with
\code{\link[latentnet]{ergmm}}.
}
\details{
Returns a performance object for a model estimates with
\code{\link[latentnet]{ergmm}}. Predictions from the supplied model are
based on the posterior mean of the parameter estimates.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
