% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistart.R
\name{multistart}
\alias{multistart}
\title{Evaluate nodes for value in invasion detection (invasion starting from each node in turn)}
\usage{
multistart(adjmat, stoch)
}
\arguments{
\item{adjmat}{adjacency matrix for evaluation}

\item{stoch}{logical var indicating whether adjacency matrix entries are fixed (1 or 0) or probabilities}
}
\description{
Evaluate nodes for value in invasion detection (invasion starting from each node in turn) using function onestart for one realization.  The output is a matrix with rows = starting nodes (introduction nodes), columns = samping nodes, and entries = number of nodes not invaded by time detected at the sampling node having started at the introduction node.
}
\examples{
Amat <- matrix(c(1,0,0,0,1,1,0,0,0,1,1,0,1,1,1,1),nrow=4,ncol=4)
multistart(adjmat=Amat, stoch=F)
sAmat <- Amat * 0.7 # each potential link has probability 0.7 of existing in one realization
multistart(adjmat=sAmat, stoch=T)
}
\keyword{prioritization}
\keyword{sampling}
