% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multsame2.R
\name{multsame2}
\alias{multsame2}
\title{Runs and summarizes multiple INA simulations with the same parameter values}
\usage{
multsame2(
  nreals2,
  usethreshman2,
  ntimesteps2,
  geocoords2,
  readgeocoords2,
  maneffdir2,
  maneffmean2,
  maneffsd2,
  maneffthresh2,
  sampeffort2,
  xrange2,
  yrange2,
  numnodes2,
  randgeo2,
  readinitinfo2,
  initinfo2,
  initinfo.dist2,
  initinfo.n2,
  initinfo.norp2,
  initinfo.p2,
  readinitbio2,
  initbio2,
  initbio.dist2,
  initbio.n2,
  initbio.norp2,
  initbio.p2,
  readseam2,
  seam2,
  seamdist2,
  seampla2,
  seamplb2,
  seamrandp2,
  readbpam2,
  bpam2,
  bpamdist2,
  bpampla2,
  bpamplb2,
  bpamrandp2,
  probadoptmean2,
  probadoptsd2,
  probadoptvec2,
  readprobadoptvec2,
  probestabmean2,
  probestabsd2,
  readprobestabvec2,
  probestabvec2,
  doplot2 = F
)
}
\arguments{
\item{nreals2}{number of realizations to be evaluated}

\item{usethreshman2}{if T, use the threshold for the management effect size estimate; thus, information is never present anywhere unless the management effect estimate exceeds the threshold}

\item{ntimesteps2}{number of time steps to be evaluated}

\item{geocoords2}{matrix of x,y coordinates of nodes, read in if readgeocoords2=T}

\item{readgeocoords2}{if T, read in geocoords2 - otherwise, generate it in each realization}

\item{maneffdir2}{if maneffdir2='decrease_estab', the management reduces the probability of establishment; if maneffdir2='increase_estab', the management reduces the probability that establishment does NOT occur}

\item{maneffmean2}{(estinfo, estab) the underlying mean change in establishment probability (as a proportion)}

\item{maneffsd2}{(estab) sd of management effect (same as maneffsd2 to function estinfo)}

\item{maneffthresh2}{(estinfo) the threshold effect size for communicating about management (if maneffthresh2 = 0 there is no threshold so communication can always occur)}

\item{sampeffort2}{(estinfo) sampling effort, where greater samping effort reduces the error in estimating the management effect}

\item{xrange2}{(genlocs) range of x coordinates}

\item{yrange2}{(genlocs) range of y coordinates}

\item{numnodes2}{(genlocs) the number of nodes}

\item{randgeo2}{(genlocs) if TRUE then locations are randomly generated}

\item{readinitinfo2}{info (setup2) if T, the initial values for the vector of starting locations for the presence of information are read in rather than generated}

\item{initinfo.dist2}{info (initvals) the pattern of locations where initial presence occurs: 'random' indicates all equally likely, 'upedge' indicates that nodes closest to the upper edge have presence, 'rightedge' indicates that nodes closest to the right edge have presence}

\item{initinfo.n2}{info (initvals) the number of initial locations for presence}

\item{initinfo.norp2}{info (initvals) 'num' indicates initial number for presence, 'prop' indicates initial proportion}

\item{initinfo.p2}{info (initvals) the proportion of initial locations for presence}

\item{readinitbio2}{bio (setup2) if T, the initial values for the vector of starting locations for the presence of the bioentity are read in rather than generated}

\item{initbio2}{bio (setup2) the vector of initial values read in if readinitbio2 == T, a vector with lenth equal to the number of nodes and entries 1s or 0s with 1s indicating the initial presence of the bioentity}

\item{initbio.dist2}{estab (initvals) the type of locations where initial presence occurs: 'random' indicates all equally likely, 'upedge' indicates that nodes closest to the upper edge have presence, 'rightedge' indicates that nodes closest to the right edge have presence}

\item{initbio.n2}{estab (initvals) the number of initial locations for presence}

\item{initbio.norp2}{estab (initvals) 'num' indicates initial number for presence, 'prop' indicates initial proportion}

\item{initbio.p2}{estab (initvals) the proportion of initial locations for presence}

\item{readseam2}{if T, the communication adjacency matrix is read in rather than being generated from the outset}

\item{seam2}{communication adjacency matrix, read in if readseam2=T}

\item{seamdist2}{(genmovnet) the function of distance used to estimate movement probability - 'random' (not related to distance) or 'powerlaw'}

\item{seampla2}{(genmovnet) power law parameter a in ad^(-b)}

\item{seamplb2}{(genmovnet) power law parameter b in ad^(-b)}

\item{seamrandp2}{(genmovnet) random case, probability of link}

\item{readbpam2}{if T, the dispersal adjacency matrix is read in rather than being generated from the outset}

\item{bpam2}{dispersal adjacency matrix, read in if readbpam2=T}

\item{bpamdist2}{(genmovnet) the function of distance used to estimate movement probability - 'random' (not related to distance) or 'powerlaw'}

\item{bpampla2}{(genmovnet) power law parameter a in ad^(-b)}

\item{bpamplb2}{(genmovnet) power law parameter b in ad^(-b)}

\item{bpamrandp2}{(genmovnet) random case, probably of link}

\item{probadoptmean2}{(makedec) mean probability of adopting management if informed}

\item{probadoptsd2}{(makedec) sd in truncated normal distribution of probability of adoption}

\item{probadoptvec2}{vector of probabilities of adoption if informed for nodes in the network (\code{readprobadoptvec2} determines whether \code{probadoptvec2} is read in to \code{makedec} or generated within \code{makedec} based on \code{probadoptmean2} and \code{probadoptsd2})}

\item{readprobadoptvec2}{if T, then a VECTOR of probability of adoption values is read in; if F, \code{probadoptmean2} and \code{probadoptsd2} are used to generate the vector}

\item{probestabmean2}{(estab) mean probability of establishment (new or CONTINUED) in absence of management}

\item{probestabsd2}{(estab) sd in truncated normal distribution for probability of establishment in absence of management}

\item{readprobestabvec2}{(estab) if T, then a vector \code{probestabvec2} is read in; otherwise the vector is generated using \code{probestabmean2} and \code{probestabsd2}}

\item{probestabvec2}{(estab) vector of probabilities of establishment (read in or generated when the function is run, depending on \code{readprobestabvec2} equal to T or F)}

\item{doplot2}{if true plots of resulting presence of information and species are generated}

\item{initinfo.s2}{info (setup2) the vector of initial values read in if readinitinfo2 == T, a vector with lenth equal to the number of nodes and entries 1s or 0s with 1s indicating the initial presence of information}
}
\description{
This function runs and summarizes multiple INA simulations with the same parameter values.  (It uses INA functions estinfo, genlocs, initvals, setup2, genmovnet, spreadstep, makedec, estab, and ntsteps2.)
}
\details{
Updated 2020-09-05
}
\examples{
x5 <- multsame2(nreals2=10, ntimesteps2=3, usethreshman2=F, readgeocoords2=T, geocoords2=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3),byrow=T,ncol=2), maneffdir2='decrease_estab', maneffmean2=0.5, maneffsd2=0.1, maneffthresh2=0.5, sampeffort2=1, xrange2=NA, yrange2=NA, numnodes2=NA, randgeo2=NA, readinitinfo2=F, initinfo.dist2='random', initinfo.n2=5, initinfo.norp2='num', initinfo.p2=NA, readinitbio2=F, initbio.dist2='upedge', initbio.n2=5, initbio.norp2='num', initbio.p2=NA, readseam2=F, seam2=NA, seamdist2='powerlaw', seampla2=1, seamplb2=1, seamrandp2=NA, readbpam2=F, bpam2=NA, bpamdist2='powerlaw', bpampla2=1, bpamplb2=1, bpamrandp2=NA, probadoptmean2=0.5, probadoptsd2=0.1, probadoptvec2=NA, readprobadoptvec2=F, probestabmean2=0.5, probestabsd2=0.1, readprobestabvec2=F, probestabvec2=NA, doplot2=F)
x3 <- multsame2(nreals2=2,ntimesteps=3, usethreshman2=F, readgeocoords2=T, geocoords2=matrix(runif(n=100)*100,byrow=T,ncol=2), maneffdir2='decrease_estab', maneffmean2=0.5, maneffsd2=0.1, maneffthresh2=0.5, sampeffort2=1, xrange2=NA, yrange2=NA, numnodes2=NA, randgeo2=NA, readinitinfo2=F, initinfo.dist2='random', initinfo.n2=5, initinfo.norp2='num', initinfo.p2=0.05, readinitbio2=F, initbio.dist2='upedge', initbio.n2=5, initbio.norp2='num', initbio.p2=0.05, readseam2=F, seam2=NA, seamdist2='powerlaw', seampla2=1, seamplb2=1, seamrandp2=NA, readbpam2=F, bpam2=NA, bpamdist2='powerlaw', bpampla2=1, bpamplb2=1, bpamrandp2=NA, probadoptmean2=0.5, probadoptsd2=c(0,1), probadoptvec2=NA, readprobadoptvec2=F, probestabmean2=0.5, probestabsd2=0.1, readprobestabvec2=F, probestabvec2=NA, doplot2=F)
}
\keyword{experiments}
\keyword{simulation}
