% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntsteps2.R
\name{ntsteps2}
\alias{ntsteps2}
\title{Changes management info status and bioentity establishment status at each node across multiple time steps}
\usage{
ntsteps2(
  nsteps,
  geocoords3n,
  infon,
  vect1cn,
  vect1dn,
  readseam3,
  seam3,
  seamdist3,
  seampla3,
  seamplb3,
  seamrandp3,
  readbpam3,
  bpam3,
  bpamdist3,
  bpampla3,
  bpamplb3,
  bpamrandp3,
  readprobadoptvec3,
  probadoptvec3,
  probadoptmean3,
  probadoptsd3,
  maneffdir3,
  maneffmean3n,
  maneffsd3n,
  readprobestabvec3,
  probestabvec3,
  probestabmean3,
  probestabsd3,
  plotmpn = F
)
}
\arguments{
\item{nsteps}{number of time steps to be evaluated}

\item{geocoords3n}{matrix of x,y coordinates of nodes}

\item{infon}{is T if estimated effect size exceeds threshold for communication (so communication can occur), F otherwise (can be evaluated in estinfo and/or setup2)}

\item{vect1cn}{com (spreadstep) status of nodes before spread}

\item{vect1dn}{disp (spreadstep) status of nodes before spread}

\item{readseam3}{if T, the communication adjacency matrix is read in rather than being generated by function genmovnet}

\item{seam3}{communication adjacency matrix, read in if readseam3=T (rows as sources and columns as sinks)}

\item{seamdist3}{com (genmovnet) the function of distance used to estimate movement probability - 'random' (not related to distance) or 'powerlaw'}

\item{seampla3}{com (genmovnet) inverse power law parameter a in ad^(-b)}

\item{seamplb3}{com (genmovnet) inverse power law parameter b in ad^(-b)}

\item{seamrandp3}{com (genmovnet) random case, probably of link}

\item{readbpam3}{if T, the dispersal adjacency matrix is read in rather than being generated by function genmovnet}

\item{bpam3}{dispersal adjacency matrix, read in if readbpam3=T (rows as sources and columns as sinks)}

\item{bpamdist3}{disp (genmovnet) the function of distance used to estimate movement probability - 'random' (not related to distance) or 'powerlaw'}

\item{bpampla3}{disp (genmovnet) inverse power law parameter a in ad^(-b)}

\item{bpamplb3}{disp (genmovnet) inverse power law parameter b in ad^(-b)}

\item{bpamrandp3}{disp (genmovnet) random case, probably of link}

\item{readprobadoptvec3}{if T, then a VECTOR of probadoptvec3 values is read in; if F, \code{probadoptmean3} and \code{probadoptsd3} are used to generate the vector}

\item{probadoptvec3}{vector of probabilities of adoption if informed for nodes in the network (\code{readprobadoptvec3} determines whether \code{probadoptvec3} is read in to \code{makedec} or generated within \code{makedec} based on \code{probadoptmean3} and \code{probadoptsd3})}

\item{probadoptmean3}{(makedec) mean probability of adopting management if informed}

\item{probadoptsd3}{(makedec) sd in truncated normal distribution of probability of adoption}

\item{maneffdir3}{if maneffdir3='decrease_estab', the management reduces the probability of establishment; if maneffdir3='increase_estab', the management reduces the probability that establishment does NOT occur}

\item{maneffmean3n}{(estab) mean management effect (proportional reduction in estabp)}

\item{maneffsd3n}{(estab) sd of management effect}

\item{readprobestabvec3}{(estab) if T, then a vector \code{probestabvec3} is read in; otherwise the vector is generated using \code{probestabmean3} and \code{probestabsd3}}

\item{probestabvec3}{(estab) vector of probabilities of establishment (read in or generated when the function is run, depending on \code{readprobestabvec3} equal to T or F)}

\item{probestabmean3}{(estab) mean probability of establishment (new or CONTINUED) in absence of management}

\item{probestabsd3}{(estab) sd in probability of establishment in truncated normal distribution}

\item{plotmpn}{if T, plots of resulting presence of information and species are generated}

\item{comvec}{vector of 1=info is present, 0=info is not present}
}
\description{
This function gives the info status and establishment status at each node after multiple time steps. Its use follows use of the function \code{setup2}, or another source of node locations and initial conditions.  (Note that communication and dispersal status are given for the beginning of the time step (rather than for the end), so the corresponding output for these has length equal to the number of time steps plus 1.  In contrast, adoption and establishment output is of length equal to the number of time steps.) It uses the following functions, as needed: genmovnet, spreadstep, makedec, and estab. It draws on output from function setup2 in terms of whether communication occurs (estinfo), generated geographic locations of nodes as needed (genloc), and the initial locations for management information and the bioentity.
}
\details{
Updated 2020-09-05
}
\examples{
x2 <- ntsteps2(nsteps=3, infon=T, geocoords3n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3),byrow=T,ncol=2), vect1cn=c(1,1,1,0,0,0), vect1dn=c(0,0,0,1,1,1), readseam3=F, seamdist3='powerlaw', seampla3=1, seamplb3=1, readbpam3=F, bpamdist3='powerlaw', bpampla3=1, bpamplb3=1, probadoptmean3=0.5, probadoptsd3=0.1, probestabmean3=0.5, probestabsd3=0.1, maneffdir3='decrease_estab', maneffmean3n=0.5, maneffsd3n=0.1,  readprobestabvec3=F, readprobadoptvec3=F, plotmpn=T)
x2z <- ntsteps2(nsteps=3, infon=T, geocoords3n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3),byrow=T,ncol=2), vect1cn=c(1,1,1,0,0,0), vect1dn=c(0,0,0,1,1,1), readseam3=F, seamdist3='powerlaw', seampla3=0.5, seamplb3=2, readbpam3=F, bpamdist3='powerlaw', bpampla3=0.5, bpamplb3=2, probadoptmean3=0.5, probadoptsd3=0.1, probestabmean3=0.5, probestabsd3=0.1, maneffdir3='decrease_estab', maneffmean3n=0.5, maneffsd3n=0.1, readprobestabvec3=F, readprobadoptvec3=F, plotmpn=T)
x3 <- ntsteps2(nsteps=2, infon=T, geocoords3n=matrix(runif(n=100)*100,byrow=T,ncol=2), vect1cn=c(rep(1,10),rep(0,40)), vect1dn=sample(c(rep(1,10),rep(0,40))), readseam3=F, seamdist3='powerlaw', seampla3=1, seamplb3=1, readbpam3=F, bpamdist3='powerlaw', bpampla3=1, bpamplb3=1, probadoptmean3=0.5, probadoptsd3=0.1, probestabmean3=0.5, probestabsd3=0.1, maneffdir3='decrease_estab', maneffmean3n=0.5, maneffsd3n=0.1, readprobestabvec3=F, readprobadoptvec3=F, plotmpn=T)
x9 <- ntsteps2(nsteps=3, infon=T, geocoords3n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3),byrow=T,ncol=2), vect1cn=c(1,1,1,0,0,0), vect1dn=c(0,0,0,1,1,1), readseam3=T, seam3=matrix(c(0,1,0,0,0,0, 0,0,1,0,0,0, 0,0,0,1,0,0, 0,0,0,0,1,0, 0,0,0,0,0,1, 1,0,0,0,0,0),byrow=T,ncol=6), readbpam3=T, bpam3=matrix(c(0,0,0,0,0,0, 0,0,1,0,0,0, 0,0,0,1,0,0, 0,0,0,0,1,0, 0,0,0,0,0,1, 1,0,0,0,0,0),byrow=T,ncol=6), probadoptmean3=0.3, probadoptsd3=0.1, probestabmean3=0.2, probestabsd3=0.1, maneffdir3='decrease_estab', maneffmean3n=0.5, maneffsd3n=0.1, readprobestabvec3=F, readprobadoptvec3=F, plotmpn=T)
x9p <- ntsteps2(nsteps=3, infon=T, geocoords3n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3),byrow=T,ncol=2), vect1cn=c(1,1,1,0,0,0), vect1dn=c(0,0,0,1,1,1), readseam3=T, seam3=matrix(c(0.5,1,0.2,0.6,0.1,0.3, 0,0,1,0,0,0, 0,0,0,1,0,0, 0,0,0,0,1,0, 0,0,0,0,0,1, 1,0,0,0,0,0),byrow=T,ncol=6), readbpam3=T, bpam3=matrix(c(0.2,0.2,0.2,0.2,0.2,0.2, 0,0,1,0,0,0, 0,0,0,1,0,0, 0,0,0,0,1,0, 0,0,0,0,0,1, 1,0,0,0,0,0),byrow=T,ncol=6), probadoptmean3=0.3, probadoptsd3=0.1, probestabmean3=0.2, probestabsd3=0.1, maneffdir3='decrease_estab', maneffmean3n=0.5, maneffsd3n=0.1, readprobestabvec3=F, readprobadoptvec3=F, plotmpn=T)
x12 <- ntsteps2(nsteps=3, infon=F, geocoords3n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3),byrow=T,ncol=2), vect1cn=c(1,1,1,0,0,0), vect1dn=c(0,0,0,1,1,1), readseam3=F, seamdist3='powerlaw', seampla3=1, seamplb3=1, readbpam3=F, bpamdist='powerlaw', bpampla3=1, bpamplb3=1, probadoptmean3=0.3, probadoptsd3=0.1, probestabmean3=0.2, probestabsd3=0.1, maneffdir3='decrease_estab', maneffmean3n=0.5, maneffsd3n=0.1, readprobestabvec3=F, readprobadoptvec3=F, plotmpn=T)
x13 <- ntsteps2(nsteps=5, infon=T, geocoords3n=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3),byrow=T,ncol=2), vect1cn=c(1,1,1,0,0,0), vect1dn=c(0,0,0,1,1,1), readseam3=F, seamdist3='powerlaw', seampla3=1, seamplb3=1, readbpam3=F, bpamdist3='powerlaw', bpampla3=1, bpamplb3=1, probadoptmean3=0.3, probadoptsd3=0.1, probestabmean3=0.2, probestabsd3=0.1, maneffdir3='decrease_estab', maneffmean3n=0.5, maneffsd3n=0.1, readprobestabvec3=F, readprobadoptvec3=F, plotmpn=T)
}
\keyword{experiments}
\keyword{simulation}
