% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_authorship_matrix}
\alias{plot_authorship_matrix}
\title{Plots 4 network graphs for either \code{authorship_matrix_1} or \code{authorship_matrix_1}.}
\usage{
plot_authorship_matrix(authorship_matrix_name, directed = FALSE)
}
\arguments{
\item{directed}{(bool) Enables or disables the \code{directed} parameter
within \code{qgraph} function.}

\item{tesc_matrix_name}{(character) The name of the authorship matrix to plot.
The name, not the matrix itself.}
}
\description{
The graphs are stored in the current working directory.
}
