% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_tesc_matrix}
\alias{plot_tesc_matrix}
\title{Plots 4 network graphs for either \code{tesc_matrix_1} or \code{tesc_matrix_3}.}
\usage{
plot_tesc_matrix(tesc_matrix_name, directed = FALSE)
}
\arguments{
\item{tesc_matrix_name}{(character) The name of the TESC matrix to plot.
The name, not the matrix itself.}

\item{directed}{(bool) Enables or disables the \code{directed} parameter
within \code{qgraph} function.}
}
\description{
The graphs are stored in the current working directory.
}
