% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentnet.R
\name{predict_response_hist}
\alias{predict_response_hist}
\title{Simulate and Plot Histogram of Outcomes from Latent Space Model}
\usage{
predict_response_hist(model, network, x, i, j, n = 1000, barplot = FALSE)
}
\arguments{
\item{model}{\code{\link[latentnet]{ergmm}} model object}

\item{network}{\code{\link[network]{network}} object from which the model
was estimated.}

\item{x}{string, the covariate outcome of interest}

\item{i}{integer, sender id}

\item{j}{integer, receiver id}

\item{n}{integer, number of simulations}

\item{barplot}{boolean, plot a barplot instead of a histogram. Useful for
discrete outcomes.}
}
\description{
Simulates and then plots outcomes between specified nodes.
}
\details{
Simulates and then plots outcomes between specified nodes from an estimated
latent space model.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
