% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ScottishSchool}
\alias{ScottishSchool}
\title{Scottish Schoolgirl Friendship Network}
\format{Two matrices and a data frame:
\describe{
\item{Girls}{Network object aggregated over the three years. Edge attribute
    'mentions' indicates the number of times, out of three, that node i
    listed j in the surveys.}
\item{Girls1}{Network object from wave 1.}
\item{Girls2}{Network object from wave 2.}
\item{Girls3}{Network object from wave 3.}
}}
\source{
\url{https://www.stats.ox.ac.uk/~snijders/siena/Glasgow_data.htm}
}
\usage{
data(ScottishSchool)
}
\description{
This is a friendship network comprised of 50 girls from a Scottish school.
The girls were surveyed over three years starting when they were about
13 years old in 1995.
}
\references{
Michell, Lynn, and Amanda Amos. "Girls, pecking order
and smoking." Social science & medicine 44, no. 12 (1997): 1861-1869.
}
\keyword{datasets}
