% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Senate}
\alias{Senate}
\title{Cosponsorship network among the 5 most liberal and 5 most conservative}
\format{A directed statnet network object with 10 nodes:
\describe{
  \item{vertex.names}{vertex attribute, name of each senator
      in the network}
  \item{ideol}{vertex attribute, DW-nominate ideology score of each senator}
  \item{dist}{network attribute, Euclidean distance in lat/lon units,
              between state capitals}
}}
\source{
\url{http://jhfowler.ucsd.edu/cosponsorship.htm}
}
\usage{
data(Senate)
}
\description{
In this directed network an edge from i to j indicates that senator i
cosponsored at least two bills sponsored by j in the 106th Congress.
}
\references{
Fowler, J. H. (2006). Connecting the Congress: A study of
    cosponsorship networks. Political Analysis, 14(4), 456-487.
}
\keyword{datasets}
