% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentnet.R
\name{simulate_edge}
\alias{simulate_edge}
\title{Simulate Outcomes from Latent Space Model}
\usage{
simulate_edge(model, n, x, i, j, seed = NULL)
}
\arguments{
\item{model}{\code{\link[latentnet]{ergmm}} model object}

\item{n}{integer, number of simulations}

\item{x}{string, the covariate outcome of interest}

\item{i}{integer, sender id}

\item{j}{integer, receiver id}

\item{seed}{integer, random seed for replicability}
}
\value{
vector of simulated values
}
\description{
Simulate edge outcomes from an estimated \code{\link[latentnet]{ergmm}} model.
}
\details{
Simulate edge outcomes from an estimated \code{\link[latentnet]{ergmm}} model.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
