% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squish.R
\name{squish}
\alias{squish}
\title{Function for plotting network on an angle}
\usage{
squish(coords, yflat = 0.5, xover = 0.5)
}
\arguments{
\item{coords}{matrix of coordinates, with each row a node, the first column the x coordinate, and the second column the y}

\item{yflat}{proportional flattening of map}

\item{xover}{proportional right movement of map}
}
\description{
This function plots a network on the angle specified, for layering plots
}
\examples{
Z1 <- matrix(c(0.5,0.8, 0.9,0.9, 0.1,0.1, 0.5,0.5), byrow=T, ncol=2)
Z1
plot(Z1, xlim=c(0,2), ylim=c(0,1))
Z2 <- squish(Z1)
plot(Z2, xlim=c(0,2), ylim=c(0,1))
}
\keyword{multilayer}
\keyword{networks}
\keyword{plotting}
