% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{str_trim}
\alias{str_trim}
\title{Internal. Trims the leading and trailing space of a string.}
\usage{
str_trim(list_authors)
}
\arguments{
\item{list_authors}{(list of vectors)}
}
\description{
Internal. Trims the leading and trailing space of a string.
}
