% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{strip_initials}
\alias{strip_initials}
\title{Internal. Strips the initials from a name. E.g., M. A. Constantin -> Constantin.}
\usage{
strip_initials(list_authors)
}
\arguments{
\item{list_authors}{(list of vectors)}
}
\description{
Internal. Strips the initials from a name. E.g., M. A. Constantin -> Constantin.
}
