% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{tesc_authors}
\alias{tesc_authors}
\title{Internal. Used to create \code{tesc_data} dataset.}
\usage{
tesc_authors()
}
\description{
Returns a data frame containing the TESC authors as seen in the document provided
by Loes. First column is the full name, second is the department.
}
