% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{tesc_matrix}
\alias{tesc_matrix}
\title{Internal. Used to build the \code{tesc_matrix_1} and \code{tesc_matrix_1} datasets.}
\usage{
tesc_matrix(authorship_data, tesc_data, verbose = T)
}
\arguments{
\item{authorship_data}{(list of vectors) Each vectors represents the TSB authors that collaborated on that paper.}

\item{tesc_data}{(data frame) A square matrix with only the authors that are TESC collaborators and the output of the \code{tesc_data} function.}

\item{verbose}{(bool)}
}
\value{
A square matrix with only the authors that are TESC collaborators and the subsequent weights indicating the way they collaborated.
}
\description{
Builds the matrix of weights only for the people that are currently part of TESC.
For each paper it checks which authors are part of the TESC and it extracts
them preserving the order. The newly extract vector for a particular
paper represents how people TESC members collaborated in that case.
The first one is considered to have collaborated with the rest
of the authors in the vector. Each such edge is awarded
one point in the matrix of weights.
}
