% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentnet.R
\name{to_sociomatrix}
\alias{to_sociomatrix}
\title{Return Weighted Adjacency Matrix}
\usage{
to_sociomatrix(network, covariate)
}
\arguments{
\item{network}{\code{\link[network]{network}} object.}

\item{covariate}{string, the covariate to extract}
}
\value{
Adjacency matrix.
}
\description{
Recover the weighted adjacency matrix for a network.
}
\details{
Recover the weighted adjacency matrix for a network.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
